Create a generic for Glyphicon+Tooltip
